backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
fastapi_mcp
google-genai>=1.31.0
websockets>=12.0
pydantic>=2.0.0